f134: integrate builder dao support

+++ contracts/src/delegate.sol
@@ -14,7 +14,7 @@ contract Delegate is DelegateEvents {
     string public constant name = "federation";
 
     /// @notice the address of the vetoer
-    address public vetoer;    
+    address public vetoer;
 
     /// @notice the address of the controlling DAO logic contract
     NounsDAOStorageV1 public ownerDAO;
@@ -40,15 +40,9 @@ contract Delegate is DelegateEvents {
      * @param execWindow_ The window in blocks that a proposal which has met quorum can be executed
      */
     constructor(address _vetoer, address nounishToken_, NounsDAOStorageV1 ownerDAO_, uint256 execWindow_) {
-        require(
-            address(ownerDAO_) != address(0),
-            "invalid owner dao address"
-        );
+        require(address(ownerDAO_) != address(0), "invalid owner dao address");
 
-        require(
-            nounishToken_ != address(0),
-            "invalid nounish NFT address"
-        );
+        require(nounishToken_ != address(0), "invalid nounish NFT address");
 
         nounishToken = NounsTokenLike(nounishToken_);
         ownerDAO = ownerDAO_;
@@ -57,15 +51,13 @@ contract Delegate is DelegateEvents {
     }
 
     function _externalProposal(NounsDAOStorageV1 eDAO, uint256 ePropID) public view returns (uint256) {
-        (,,,,,, uint256 ePropEndBlock,,,,,,) = eDAO.proposals(
-            ePropID
-        );
+        (,,,,,, uint256 ePropEndBlock,,,,,,) = eDAO.proposals(ePropID);
 
         return ePropEndBlock;
     }
 
     function _alreadyProposed(address eDAO, uint256 ePropID) public view returns (bool) {
-        for (uint i=1; i <= proposalCount; i++){
+        for (uint256 i = 1; i <= proposalCount; i++) {
             if (proposals[i].eDAO == eDAO && proposals[i].eID == ePropID) {
                 return true;
             }
@@ -81,37 +73,29 @@ contract Delegate is DelegateEvents {
      * @return Proposal id of internal delegate action
      */
     function propose(NounsDAOStorageV1 eDAO, uint256 ePropID) public returns (uint256) {
-        require(
-            nounishToken.getPriorVotes(msg.sender, block.number - 1) > 0,
-            "representation required to start a vote"
-        );
+        require(nounishToken.getPriorVotes(msg.sender, block.number - 1) > 0, "representation required to start a vote");
 
-        require(
-            address(eDAO) != address(0),
-            "external DAO address is not valid"
-        );
+        require(address(eDAO) != address(0), "external DAO address is not valid");
 
-        require (
-            !_alreadyProposed(address(eDAO), ePropID),
-            "proposal already proposed"
-        );
+        require(!_alreadyProposed(address(eDAO), ePropID), "proposal already proposed");
 
         // this delegate must have representation before voting can be started
         try eDAO.nouns().getPriorVotes(address(this), block.number - 1) returns (uint96 votes) {
-            require(votes > 0, "delegate does not have external DAO representation");   
-        } catch(bytes memory) {
+            require(votes > 0, "delegate does not have external DAO representation");
+        } catch (bytes memory) {
             revert("checking delegate representation on external DAO failed");
-        }        
+        }
 
         // check when external proposal ends
         uint256 ePropEndBlock;
         try this._externalProposal(eDAO, ePropID) returns (uint256 endBlock) {
             ePropEndBlock = endBlock;
-        } catch(bytes memory) {
+        } catch (bytes memory) {
             revert(
-                string.concat("invalid external proposal id: ", 
-                    Strings.toString(ePropID), 
-                    " for external DAO: ", 
+                string.concat(
+                    "invalid external proposal id: ",
+                    Strings.toString(ePropID),
+                    " for external DAO: ",
                     Strings.toHexString(address(eDAO))
                 )
             );
@@ -125,10 +109,7 @@ contract Delegate is DelegateEvents {
         newProposal.eID = ePropID;
         newProposal.eDAO = address(eDAO);
         newProposal.proposer = msg.sender;
-        newProposal.quorumVotes = bps2Uint(
-            ownerDAO.quorumVotesBPS(),
-            nounishToken.totalSupply()
-        );
+        newProposal.quorumVotes = bps2Uint(ownerDAO.quorumVotesBPS(), nounishToken.totalSupply());
 
         /// @notice immediately open proposal for voting
         newProposal.startBlock = block.number;
@@ -149,7 +130,7 @@ contract Delegate is DelegateEvents {
             newProposal.startBlock,
             newProposal.endBlock,
             newProposal.quorumVotes
-        );
+            );
 
         return newProposal.id;
     }
@@ -160,22 +141,16 @@ contract Delegate is DelegateEvents {
      * @dev This function ensures that the proposal has reached quorum through a result check
      */
     function execute(uint256 proposalId) external {
-        require(
-            state(proposalId) == ProposalState.Active,
-            "proposal can only be executed if it is active"
-        );
+        require(state(proposalId) == ProposalState.Active, "proposal can only be executed if it is active");
 
         ProposalResult r = result(proposalId);
-        require(
-            r != ProposalResult.Undecided,
-            "proposal result cannot be undecided"
-        );
+        require(r != ProposalResult.Undecided, "proposal result cannot be undecided");
 
         DelegateAction storage proposal = proposals[proposalId];
         proposal.executed = true;
 
         require(
-            block.number >= proposal.endBlock-execWindow,
+            block.number >= proposal.endBlock - execWindow,
             "proposal can only be executed if it is within the execution window"
         );
 
@@ -199,34 +174,24 @@ contract Delegate is DelegateEvents {
      */
     function veto(uint256 proposalId) external {
         require(vetoer != address(0), "veto power burned");
-        
+
         require(msg.sender == vetoer, "caller not vetoer");
-        
-        require(
-            state(proposalId) != ProposalState.Executed,
-            "cannot veto executed proposal"
-        );
+
+        require(state(proposalId) != ProposalState.Executed, "cannot veto executed proposal");
 
         DelegateAction storage proposal = proposals[proposalId];
         proposal.vetoed = true;
 
         emit ProposalVetoed(proposalId);
-    }    
+    }
 
     /**
      * @notice Cast a vote for a proposal with an optional reason
      * @param proposalId The id of the proposal to vote on
      * @param support The support value for the vote. 0=against, 1=for, 2=abstain
      */
-    function castVote(
-        uint256 proposalId,
-        uint8 support,
-        string calldata reason
-    ) external {
-        require(
-            state(proposalId) == ProposalState.Active,
-            "voting is closed"
-        );
+    function castVote(uint256 proposalId, uint8 support, string calldata reason) external {
+        require(state(proposalId) == ProposalState.Active, "voting is closed");
 
         require(support <= 2, "invalid vote type");
 
@@ -237,10 +202,7 @@ contract Delegate is DelegateEvents {
 
         Receipt storage receipt = proposal.receipts[msg.sender];
 
-        require(
-            receipt.hasVoted == false,
-            "already voted"
-        );
+        require(receipt.hasVoted == false, "already voted");
 
         if (support == 0) {
             proposal.againstVotes = proposal.againstVotes + votes;
@@ -254,13 +216,7 @@ contract Delegate is DelegateEvents {
         receipt.support = support;
         receipt.votes = votes;
 
-        emit VoteCast(
-            msg.sender,
-            proposalId,
-            support,
-            votes,
-            reason
-        );
+        emit VoteCast(msg.sender, proposalId, support, votes, reason);
     }
 
     /**
@@ -304,7 +260,7 @@ contract Delegate is DelegateEvents {
 
         DelegateAction storage proposal = proposals[proposalId];
 
-        uint256 totalVotes = proposal.forVotes + proposal.againstVotes + proposal.abstainVotes ;
+        uint256 totalVotes = proposal.forVotes + proposal.againstVotes + proposal.abstainVotes;
         if (totalVotes < proposal.quorumVotes) {
             return ProposalResult.Undecided;
         }
@@ -328,7 +284,7 @@ contract Delegate is DelegateEvents {
      * @notice Changes proposal exec window
      * @dev function for updating the exec window of a proposal
      */
-    function _setExecWindow(uint newExecWindow) public {
+    function _setExecWindow(uint256 newExecWindow) public {
         require(msg.sender == address(ownerDAO), "only owner can change exec window");
 
         emit NewExecWindow(execWindow, newExecWindow);
